Night order formatting and generate_script.py changes
The Other Nights text now stretches to the bottom of the page based on the value of the teensyville boolean, but if it's false the text will stretch to the bottom of the first night text. This was extremely annoying to get working right

Old .tex files are not compatibile with the new script.cls

generate_script.py has been updated to match the script.cls changes. It now handles _meta properly, and has less processing to do for tables as that's been shifted to script.cls.

Also generate_script.py now takes the full filename of the script including the .json as its argument.